requirements.txt
# AI
llama-index==0.12.17
llama-index-llms-openai==0.3.10
llama-index-llms-deepseek==0.1.1
llama-index-llms-azure-openai==0.3.0
llama-index-llms-azure-inference==0.3.0
llama-index-tools-azure-code-interpreter==0.4.0
# Server
quart==0.20.0
gunicorn==23.0.0
uvicorn==0.34.0
# Bot
azure-identity==1.20.0
botbuilder-core==4.16.2
botbuilder-ai 
botbuilder-schema 
botbuilder-integration-aiohttp
# Tools & helpers
httpx==0.23.2
python-docx==1.1.2